pyproject.toml
[project]
name = "poker-solver"
version = "0.1.0"
description = "Heads-Up No-Limit Hold'em solver using Monte Carlo CFR"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Matteo Pesenti" }
]
dependencies = [
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scipy>=1.15.3",
    "scipy-stubs>=1.15.3.0",
    "scikit-learn==1.7.2",
    "numba>=0.63.1",
    "xxhash>=3.6.0",
    "questionary>=2.1.1",
    "zarr<3",
    "numcodecs<0.16",
    "eval7>=0.1.10",
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "pydantic>=2.0",
]

[project.scripts]
poker-solver = "src.interfaces.cli.app:main"

[project.optional-dependencies]
# Visualization and performance profiling
optimization = [
    "matplotlib>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"*.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "timeout: test timeout in seconds",
]
addopts = "-v"
timeout = 5
timeout_func_only = true

[dependency-groups]
# Development, testing, and code quality tools
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
    "ty",
    "pre-commit>=4.5.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-tqdm>=4.67.0.20250809",
    "pytest-timeout>=2.4.0",
    "httpx>=0.28.0",
]

[tool.uv]
preview = true

[tool.uv.extra-build-dependencies]
eval7 = ["Cython"]
